+++ CavrnusSdk/API/CavrnusAuth.cs
@@ -1,9 +1,0 @@
-using Collab.Proxy.Comm.RestApi;
-
-namespace CavrnusSdk
-{
-	public static class CavrnusAuth
-	{
-		public static RestApiEndpoint Endpoint;
-	}
-}