+++ CavrnusSdk/API/CavrnusSpaceJoinEvent.cs
@@ -23,6 +23,8 @@ namespace CavrnusSdk
 
 		public static void InvokeSpaceJoin(CavrnusSpaceConnection cavrnusSpace)
 		{
+			_currentLoadingSpace = null;
+
 			_currentCavrnusSpace = cavrnusSpace;
 			foreach (var joinAction in onSpaceJoinActions) joinAction?.Invoke(cavrnusSpace);
 			onSpaceJoinActions.Clear();
@@ -33,6 +35,27 @@ namespace CavrnusSdk
 			if (_currentCavrnusSpace != null) _currentCavrnusSpace.Dispose();
 
 			_currentCavrnusSpace = null;
+		}
+
+		public static string CurrentLoadingSpace => _currentLoadingSpace;
+		private static string _currentLoadingSpace;
+
+		private static List<Action<string>> onSpaceLoadingActions =
+			new List<Action<string>>();
+
+		public static void OnSpaceLoading(Action<string> a)
+		{
+			if (_currentCavrnusSpace != null)
+				a?.Invoke(_currentLoadingSpace);
+			else
+				onSpaceLoadingActions.Add(a);
+		}
+
+		public static void InvokeSpaceLoading(string space)
+		{
+			_currentLoadingSpace = space;
+			foreach (var loadingAction in onSpaceLoadingActions) loadingAction?.Invoke(space);
+			onSpaceLoadingActions.Clear();
 		}
 	}
 }