+++ Samples~/Core/Scripts/UI/LoginPanel.cs
@@ -10,23 +10,13 @@ namespace CavrnusSdk
 		[SerializeField] private TMP_InputField email;
 		[SerializeField] private TMP_InputField password;
 
-		[SerializeField] private GameObject spacePickerPrefab;
-
-		void Start() { CavrnusHelpers.Setup(); }
-
-		public async void JoinCavrnusSpace()
+		public async void Authenticate()
 		{
 			try { await CavrnusHelpers.Authenticate(server.text, email.text, password.text); }
 			catch (Exception e) {
 				Debug.Log(e.ToString());
 				return;
 			}
-
-			var foundSpaces = await CavrnusHelpers.GetAllAvailableSpaces();
-
-			GameObject.Instantiate(spacePickerPrefab, transform.parent).GetComponent<SpacePicker>().Setup(foundSpaces);
-
-			GameObject.Destroy(gameObject);
 		}
 	}
 }